added confirmation pop ups and not allow to add book to booklist if data is missing

+++ library.js
@@ -3,7 +3,6 @@ let modal = document.querySelector(".modal-screen");
 let outisdeModal = document.querySelector(".modal-screen");
 let modalBox = document.querySelector(".modal");
 let openModal = document.getElementById("add-book");
-let addBookButton = document.querySelector(".submit");
 let closeModal = document.getElementById("close");
 
 let bookText = document.getElementById("book-title");
@@ -11,6 +10,7 @@ let authorText = document.getElementById("author");
 let numPagesText = document.getElementById("num-pages");
 let statusText = document.getElementById("status");
 
+let addBookButton = document.querySelector(".submit");
 let cardGrid = document.querySelector(".card-grid");
 
 let bookList = JSON.parse(localStorage.getItem("bookList")) || [];
@@ -22,20 +22,25 @@ removeBook();
 
 console.log(bookList);
 
+//functions
 function removeBook(){
     let removeBookButton = document.querySelectorAll(".remove-book");
     removeBookButton.forEach ((remove) =>{
         remove.addEventListener("click", (event)=>{
-            let book = event.target.closest(".card");
-            let index = parseInt(book.getAttribute("data-index"));
+            let confirmMessage = "Are you Sure you Want to Delete this Book?"
+            if(confirm(confirmMessage) == true){
+                let book = event.target.closest(".card");
+                let index = parseInt(book.getAttribute("data-index"));
 
-            bookList.splice(index, 1);
-            localStorage.setItem("bookList", JSON.stringify(bookList));
-            addBookCard();
-            removeBook()
+                bookList.splice(index, 1);
+                localStorage.setItem("bookList", JSON.stringify(bookList));
+                addBookCard();
+                removeBook()
+            }
         })
     })
 }
+
 function openAndCloseModal(){
     openModal.addEventListener("click", () =>{
     modal.classList.add("hidden");
@@ -59,12 +64,17 @@ function getBookInfo(){
         let numOfPages = numPagesText.value;
         let readingStatus = statusText.value;
 
+        if(!bookText || !bookAuthor || !numOfPages || !readingStatus){
+            alert("Please Fill Out the Necessary Fields First");
+            return;
+        }
         let book = new Book(bookTitle, bookAuthor, numOfPages, readingStatus);
         bookList.push(book);
 
         localStorage.setItem("bookList", JSON.stringify(bookList));
         console.log(bookList);
-    })
+        addBookCard();
+    });
 }
 
 function Book(title, author, numberOfPages, status){
